third_party/blink/web_tests/external/wpt/html/dom/elements/global-attributes/dir-auto-dynamic-changes.window-expected.txt

This is a testharness.js-based test.
[FAIL] dir=auto slot is updated if input stops being auto-directionality form-associated
  assert_equals: expected "rtl" but got "ltr"
Harness: the test ran to completion.
